implement stimulus_presentations
flake8

store all running signals in nwb file

get_running_speed returns a tuple, and get_running_data_df is the whole running data dataframe.  removed frames

Add metadata to top level of NWB file and start get_running_data_df method

Moved RunningSpeed up a level and reusing in behavior.  This makes behavior py3 only

using processing module instead of analysis to store running_df

fix running speed issues from resulting from refactoring that didnt show up because of conftest missing

added schemas

abstract base class for behavior api

rework api to not accept ophys_session_id as arg; add test fixture for running_data_df

tests passing. for data_df read api

WIP

refactored together the get_running_speed code

review feedback refactor together add_running_data_df_to_nwbfile to use add_running_speed_to_nwbfile

get_stimulus_template tot get_stimulus_templates

store all running signals in nwb file

get_running_speed returns a tuple, and get_running_data_df is the whole running data dataframe.  removed frames

Add metadata to top level of NWB file and start get_running_data_df method

Moved RunningSpeed up a level and reusing in behavior.  This makes behavior py3 only

using processing module instead of analysis to store running_df

fix running speed issues from resulting from refactoring that didnt show up because of conftest missing

added schemas

abstract base class for behavior api

rework api to not accept ophys_session_id as arg; add test fixture for running_data_df

tests passing. for data_df read api

WIP

refactored together the get_running_speed code

adds add_image_template method

WIP

test get_stimulus_templates api method

Addes stimulus_presentations for behavior session
