+++ allensdk/internal/api/behavior_ophys_api.py
@@ -8,7 +8,7 @@ from allensdk.api.cache import memoize
 from allensdk.internal.api.ophys_lims_api import OphysLimsApi
 from allensdk.brain_observatory.behavior.sync import get_sync_data, get_stimulus_rebase_function
 from allensdk.brain_observatory.behavior.roi_processing import get_roi_metrics
-from allensdk.brain_observatory.behavior.stimulus_processing import get_stimtable, get_stimulus_templates, get_stimulus_metadata
+from allensdk.brain_observatory.behavior.stimulus_processing import get_stimulus_presentations, get_stimulus_templates, get_stimulus_metadata
 from allensdk.brain_observatory.behavior.metadata_processing import get_task_parameters
 from allensdk.brain_observatory.behavior.running_processing import get_running_df
 from allensdk.brain_observatory.behavior.rewards_processing import get_rewards
@@ -122,12 +122,12 @@ class BehaviorOphysLimsApi(OphysLimsApi):
         return RunningSpeed(timestamps=running_data_df.index.values,
                             values=running_data_df.speed.values)
 
-    @memoize    
-    def get_stimulus_table(self, ophys_experiment_id=None, use_acq_trigger=False):
+    @memoize
+    def get_stimulus_presentations(self, ophys_experiment_id=None, use_acq_trigger=False):
         stimulus_timestamps = self.get_stimulus_timestamps(ophys_experiment_id=ophys_experiment_id, use_acq_trigger=use_acq_trigger)
         behavior_stimulus_file = self.get_behavior_stimulus_file(ophys_experiment_id=ophys_experiment_id)
         data = pd.read_pickle(behavior_stimulus_file)
-        return get_stimtable(data, stimulus_timestamps)
+        return get_stimulus_presentations(data, stimulus_timestamps)
 
 
     @memoize
